Deduction manipulation editor

+++ lib/vac_engine_web/live/editor_live/deduction_inspector_component.ex
@@ -2,93 +2,120 @@ defmodule VacEngineWeb.EditorLive.DeductionInspectorComponent do
   use VacEngineWeb, :live_component
 
   import VacEngine.PipeHelpers
+  alias Ecto.Changeset
   alias VacEngine.Processor
-  alias VacEngine.Processor.Blueprint
   alias VacEngineWeb.EditorLive.DeductionListComponent
+  alias VacEngine.EnumHelpers
 
   @impl true
   def mount(socket) do
     socket
-    |> assign(blueprint: nil, on_update: nil, changeset: nil)
+    |> assign(blueprint: nil, on_update: nil, changeset: nil, inspector: nil)
     |> select(nil)
     |> ok()
   end
 
   @impl true
-  def update(%{action: {:select_path, path}}, socket) do
-    {:ok, select(socket, path)}
+  def update(%{action: {:select, selection}}, socket) do
+    {:ok, select(socket, selection)}
   end
 
   @impl true
   def update(assigns, socket) do
     socket
     |> assign(assigns)
+    |> select()
     |> on_update()
     |> ok()
   end
 
   @impl true
   def handle_event(
-        "add_deduction",
+        "new_deduction",
         _,
-        %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-          socket
+        %{assigns: %{blueprint: blueprint}} = socket
+      ) do
+    can!(socket, :write, blueprint)
+    changeset = Changeset.cast({%{}, %{variable: :string}}, %{}, [])
+
+    vars =
+      blueprint.variable_path_index
+      |> Enum.map(fn {path, _var} ->
+        path = Enum.join(path, ".")
+        {path, path}
+      end)
+      |> Enum.sort()
+
+    socket
+    |> assign(changeset: changeset, inspector: :new_deduction, variables: vars)
+    |> pair(:noreply)
+  end
+
+  @impl true
+  def handle_event(
+        "add_deduction",
+        %{"deduction" => %{"variable" => variable_path}},
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
       ) do
     can!(socket, :write, blueprint)
 
     attrs =
-      case selected_path do
-        [:deductions, idx | _] -> %{position: idx + 1}
+      case selection do
+        %{deduction: %{position: idx}} -> %{position: idx + 1}
         _ -> %{}
       end
+      |> Map.merge(%{
+        branches: [%{}],
+        columns: [
+          %{
+            type: :assignment,
+            variable: variable_path |> String.split(".")
+          }
+        ]
+      })
 
     {:ok, deduction} = Processor.create_deduction(blueprint, attrs)
 
-    new_path = [:deductions, deduction.position]
-
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
+    selection = %{deduction: deduction}
 
     send(self(), :reload_blueprint)
 
-    {:noreply,
-     assign(socket,
-       on_update: {:scroll_to, ["deduction.#{deduction.id}"]}
-     )
-     |> select(new_path)}
+    socket
+    |> assign(
+      selection: selection,
+      on_update: {:scroll_to, ["deduction.#{deduction.id}"]}
+    )
+    |> pair(:noreply)
   end
 
   @impl true
   def handle_event(
         "delete_deduction",
         _,
-        %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-          socket
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
       ) do
     can!(socket, :write, blueprint)
 
-    [:deductions, idx | _] = selected_path
-
-    deduction = blueprint.deductions |> Enum.at(idx)
+    %{deduction: deduction} = selection
 
     {:ok, _deduction} = Processor.delete_deduction(deduction)
 
-    new_path =
-      if idx > 0 do
-        [:deductions, idx - 1]
+    idx =
+      if deduction.position == 0 do
+        1
+      else
+        deduction.position - 1
+      end
+
+    selection =
+      if blueprint.deductions |> Enum.count() > 1 do
+        %{deduction: blueprint.deductions |> Enum.at(idx)}
       else
         nil
       end
 
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
-
     send(self(), :reload_blueprint)
-    {:noreply, select(socket, new_path)}
+    {:noreply, assign(socket, selection: selection)}
   end
 
   @impl true
@@ -113,72 +140,61 @@ defmodule VacEngineWeb.EditorLive.DeductionInspectorComponent do
   def handle_event(
         "add_branch",
         _,
-        %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-          socket
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
       ) do
     can!(socket, :write, blueprint)
 
-    {deduction_idx, attrs} =
-      case selected_path do
-        [:deductions, deduction_idx, :branches, idx | _] -> {deduction_idx, %{position: idx + 1}}
-        [:deductions, deduction_idx | _] -> {deduction_idx, %{}}
+    {deduction, attrs} =
+      case selection do
+        %{deduction: deduction, branch: %{position: idx}} ->
+          {deduction, %{position: idx + 1}}
+
+        %{deduction: deduction} ->
+          {deduction, %{}}
       end
 
-    deduction = blueprint.deductions |> Enum.at(deduction_idx)
     {:ok, branch} = Processor.create_branch(deduction, attrs)
 
-    new_path = [:deductions, deduction_idx, :branches, branch.position]
-
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
+    selection =
+      if Map.get(selection, :column) do
+        Map.put(selection, :branch, branch)
+      else
+        %{deduction: deduction}
+      end
 
     send(self(), :reload_blueprint)
 
-    {:noreply,
-     socket
-     |> select(new_path)}
+    {:noreply, assign(socket, selection: selection)}
   end
 
   @impl true
   def handle_event(
         "delete_branch",
         _,
-        %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-          socket
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
       ) do
     can!(socket, :write, blueprint)
 
-    [:deductions, deduction_idx, :branches, idx | _] = selected_path
+    %{deduction: deduction, branch: branch} = selection
 
-    deduction = blueprint.deductions |> Enum.at(deduction_idx)
-    branch = deduction.branches |> Enum.at(idx)
+    if deduction.branches |> Enum.count() < 2 do
+      raise "cannot delete last branch"
+    end
 
     {:ok, _branch} = Processor.delete_branch(branch)
 
-    new_path =
-      cond do
-        idx > 0 ->
-          [:deductions, deduction_idx, :branches, idx - 1]
-
-        deduction.branches |> Enum.count() > 1 ->
-          [:deductions, deduction_idx, :branches, 0]
-
-        true ->
-          [:deductions, deduction_idx]
+    idx =
+      if branch.position == 0 do
+        1
+      else
+        branch.position - 1
       end
 
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
+    selection = Map.put(selection, :branch, deduction.branches |> Enum.at(idx))
 
     send(self(), :reload_blueprint)
 
-    {:noreply,
-     socket
-     |> select(new_path)}
+    {:noreply, assign(socket, selection: selection)}
   end
 
   @impl true
@@ -199,56 +215,186 @@ defmodule VacEngineWeb.EditorLive.DeductionInspectorComponent do
     move_branch(socket, 1)
   end
 
+  @impl true
+  def handle_event(
+        "new_column",
+        _,
+        %{assigns: %{blueprint: blueprint}} = socket
+      ) do
+    can!(socket, :write, blueprint)
+
+    changeset =
+      Changeset.cast({%{}, %{variable: :string, type: :string}}, %{}, [])
+
+    vars =
+      blueprint.variable_path_index
+      |> Enum.map(fn {path, _var} ->
+        path = Enum.join(path, ".")
+        {path, path}
+      end)
+      |> Enum.sort()
+
+    socket
+    |> assign(changeset: changeset, inspector: :new_column, variables: vars)
+    |> pair(:noreply)
+  end
+
+  @impl true
+  def handle_event(
+        "add_column",
+        %{"column" => %{"variable" => variable_path, "type" => type}},
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
+      ) do
+    can!(socket, :write, blueprint)
+
+    {deduction, position} =
+      case selection do
+        %{deduction: deduction, column: %{position: idx}} ->
+          {deduction, idx + 1}
+
+        %{deduction: deduction} ->
+          {deduction, 0}
+      end
+
+    attrs = %{
+      position: position,
+      variable: variable_path |> String.split("."),
+      type: type
+    }
+
+    {:ok, column} = Processor.create_column(blueprint, deduction, attrs)
+
+    selection =
+      if Map.get(selection, :branch) do
+        Map.put(selection, :column, column)
+      else
+        %{deduction: deduction}
+      end
+
+    send(self(), :reload_blueprint)
+
+    {:noreply, assign(socket, selection: selection)}
+  end
+
+  @impl true
+  def handle_event(
+        "delete_column",
+        _,
+        %{assigns: %{blueprint: blueprint, selection: selection}} = socket
+      ) do
+    can!(socket, :write, blueprint)
+
+    %{deduction: deduction, column: column} = selection
+
+    assignment_count =
+      deduction.columns
+      |> Enum.filter(&(&1.type == :assignment))
+      |> Enum.count()
+
+    can_delete =
+      if column.type == :assignment do
+        assignment_count > 1
+      else
+        true
+      end
+
+    if !can_delete do
+      raise "cannot delete last assignment column"
+    end
+
+    {:ok, _column} = Processor.delete_column(column)
+
+    idx =
+      if column.position == 0 do
+        1
+      else
+        column.position - 1
+      end
+
+    selection = Map.put(selection, :column, deduction.columns |> Enum.at(idx))
+
+    send(self(), :reload_blueprint)
+
+    {:noreply, assign(socket, selection: selection)}
+  end
+
+  @impl true
+  def handle_event(
+        "move_column_left",
+        _,
+        socket
+      ) do
+    move_column(socket, -1)
+  end
+
+  @impl true
+  def handle_event(
+        "move_column_right",
+        _,
+        socket
+      ) do
+    move_column(socket, 1)
+  end
+
+  @impl true
+  def handle_event(
+        "cancel",
+        _,
+        %{assigns: %{selection: selection}} = socket
+      ) do
+    {:noreply, socket |> select(selection)}
+  end
+
   defp move_deduction(
-         %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-           socket,
+         %{assigns: %{blueprint: blueprint, selection: selection}} = socket,
          offset
        ) do
     can!(socket, :write, blueprint)
 
-    [:deductions, idx | _] = selected_path
-    new_pos = idx + offset
+    %{deduction: deduction} = selection
+    new_pos = deduction.position + offset
 
-    deduction = blueprint.deductions |> Enum.at(idx)
-
-    {:ok, _deduction} =
+    {:ok, deduction} =
       Processor.update_deduction(deduction, %{position: new_pos})
 
-    new_path = [:deductions, new_pos]
-
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
+    selection = %{selection | deduction: deduction}
 
     send(self(), :reload_blueprint)
-    {:noreply, select(socket, new_path)}
+    {:noreply, assign(socket, selection: selection)}
   end
 
   defp move_branch(
-         %{assigns: %{blueprint: blueprint, selected_path: selected_path}} =
-           socket,
+         %{assigns: %{blueprint: blueprint, selection: selection}} = socket,
          offset
        ) do
     can!(socket, :write, blueprint)
 
-    [:deductions, deduction_idx, :branches, idx | _] = selected_path
+    %{branch: %{position: idx} = branch} = selection
     new_pos = idx + offset
 
-    deduction = blueprint.deductions |> Enum.at(deduction_idx)
-    branch = deduction.branches |> Enum.at(idx)
+    {:ok, branch} = Processor.update_branch(branch, %{position: new_pos})
 
-    {:ok, _deduction} = Processor.update_branch(branch, %{position: new_pos})
-
-    new_path = [:deductions, deduction_idx, :branches, new_pos]
-
-    send_update(DeductionListComponent,
-      id: "deduction_list",
-      action: {:select_path, new_path}
-    )
+    selection = %{selection | branch: branch}
 
     send(self(), :reload_blueprint)
-    {:noreply, select(socket, new_path)}
+    {:noreply, assign(socket, selection: selection)}
+  end
+
+  defp move_column(
+         %{assigns: %{blueprint: blueprint, selection: selection}} = socket,
+         offset
+       ) do
+    can!(socket, :write, blueprint)
+
+    %{column: %{position: idx} = column} = selection
+    new_pos = idx + offset
+
+    {:ok, column} = Processor.update_column(column, %{position: new_pos})
+
+    selection = %{selection | column: column}
+
+    send(self(), :reload_blueprint)
+    {:noreply, assign(socket, selection: selection)}
   end
 
   defp on_update(%{assigns: %{on_update: {f, args}}} = socket)
@@ -267,49 +413,121 @@ defmodule VacEngineWeb.EditorLive.DeductionInspectorComponent do
     })
   end
 
-  defp select(socket, path) do
+  def select(socket, selection \\ :reselect) do
     blueprint = socket.assigns.blueprint
 
-    {idx, branch_idx} =
-      case path do
-        [_, idx, _, branch_idx | _] -> {idx, branch_idx}
-        [_, idx | _] -> {idx, nil}
-        _ -> {nil, nil}
+    {selection, reselect} =
+      if selection == :reselect do
+        {socket.assigns.selection, true}
+      else
+        {selection, false}
       end
 
-    count =
+    {
+      deduction,
+      column,
+      branch
+    } =
+      if is_map(selection) and not is_nil(blueprint) do
+        {
+          Map.get(selection, :deduction),
+          Map.get(selection, :column),
+          Map.get(selection, :branch)
+        }
+      else
+        {nil, nil, nil}
+      end
+
+    deduction =
+      if deduction do
+        blueprint.deductions |> EnumHelpers.find_by(:id, deduction.id)
+      else
+        nil
+      end
+
+    branch =
+      if not is_nil(deduction) and not is_nil(branch) do
+        deduction.branches |> EnumHelpers.find_by(:id, branch.id)
+      else
+        nil
+      end
+
+    column =
+      if not is_nil(deduction) and not is_nil(column) do
+        deduction.columns |> EnumHelpers.find_by(:id, column.id)
+      else
+        nil
+      end
+
+    selection = %{deduction: deduction, branch: branch, column: column}
+
+    deduction_idx =
+      case deduction do
+        nil -> nil
+        d -> d.position
+      end
+
+    branch_idx =
+      case branch do
+        nil ->
+          nil
+
+        b ->
+          b.position
+      end
+
+    deduction_count =
       case blueprint do
         nil -> 0
         br -> br.deductions |> Enum.count()
       end
 
-    branch_count =
-      case idx do
+    {branch_count, column_count, assignment_count, split_position} =
+      case deduction do
         nil ->
-          0
+          {0, 0, 0, 0}
 
-        n ->
-          blueprint.deductions
-          |> Enum.at(n)
-          |> case do
-            nil ->
-              0
+        d ->
+          assignments = d.columns |> Enum.filter(&(&1.type == :assignment))
 
-            d ->
-              Enum.count(d.branches)
-          end
+          {Enum.count(d.branches), Enum.count(d.columns),
+           assignments |> Enum.count(), Enum.at(assignments, 0).position}
+      end
+
+    if reselect do
+      send_update(DeductionListComponent,
+        id: "deduction_list",
+        action: {:select, selection}
+      )
+    end
+
+    {can_delete_column, can_move_left_column, can_move_right_column} =
+      case column do
+        %{type: :assignment, position: pos} ->
+          {assignment_count > 1, pos > split_position, pos < column_count - 1}
+
+        %{type: :condition, position: pos} ->
+          {true, pos > 0, pos < split_position - 1}
+
+        _ ->
+          {false, false, false}
       end
 
     assign(socket,
-      selected_path: path,
+      inspector: nil,
+      selection: selection,
       can_add_deduction?: true,
-      can_move_up_deduction?: idx && idx > 0,
-      can_move_down_deduction?: idx < count - 1,
-      can_delete_deduction?: idx != nil,
-      can_add_branch?: idx != nil,
+      can_move_up_deduction?: deduction_idx && deduction_idx > 0,
+      can_move_down_deduction?: deduction_idx < deduction_count - 1,
+      can_delete_deduction?: deduction_idx != nil,
+      can_add_branch?: deduction_idx != nil,
       can_move_up_branch?: branch_idx && branch_idx > 0,
       can_move_down_branch?: branch_idx < branch_count - 1,
-      can_delete_branch?: branch_idx != nil
+      can_delete_branch?: branch_idx != nil && branch_count > 1,
+      can_add_column?: deduction_idx != nil,
+      can_move_left_column?: can_move_left_column,
+      can_move_right_column?: can_move_right_column,
+      can_delete_column?: can_delete_column
     )
   end
 end
